completed funnel refactor idea #1
